requirements.txt
feedparser>=6.0
requests>=2.31
beautifulsoup4>=4.12
python-dateutil>=2.9
pydantic>=2.7
PyYAML>=6.0.1
langgraph>=0.2
tenacity>=8.2
sendgrid>=6.11
openai>=1.40
python-dotenv>=1.0